// src/app/shared/models/linea-base.ts
export interface LineaBase {
  idLineaBase?: number;
  idTipoDuracion: number;
  idObra?: number;
  fechaInicial: string;
  fechaFinal: string;
  nombre: string;
  noLineaBase?: number;
  idUsuarioModifico?: number;
  porcentaje: number;
  estatus?: number;
  tipoDuracion?: any;
  // idUsuarioCreo?: number;
}